🚀 [Feat]: study form, patch study form 분리
