+++ src/features/create-study/form/CreateStudyForm3.tsx
@@ -1,4 +1,4 @@
-import { useParams, useRouter } from "next/navigation";
+import { useRouter } from "next/navigation";
 import { FormProvider, useFormContext } from "react-hook-form";
 
 import Badge from "@/shared/components/atoms/Badge";
@@ -9,16 +9,14 @@ import StudyActions from "@/features/create-study/ui/StudyActions";
 import StudyApply from "@/features/create-study/ui/StudyApply";
 import useModal from "@/shared/hooks/useModal";
 
-import { patchStudy } from "@/entities/study/studies/api/studies.api";
-import { EditStudyForm } from "@/entities/study/studies/api/studies.api.type";
+import { postStudy } from "@/entities/study/studies/api/studies.api";
+import { StudyForm } from "@/entities/study/studies/api/studies.api.type";
 import URL from "@/shared/constants/URL";
 
 const CreateStudyForm3 = () => {
-  const params = useParams();
-  const id = params.id as string;
-  const methods = useFormContext<EditStudyForm>();
+  const methods = useFormContext<StudyForm>();
   const router = useRouter();
-  const { watch } = useFormContext<EditStudyForm>();
+  const { watch } = useFormContext<StudyForm>();
   const formData = watch();
 
   const { isModalOpenState, openModal, closeModal } = useModal();
@@ -32,9 +30,8 @@ const CreateStudyForm3 = () => {
   };
 
   const handleClickCreateButton = async () => {
-    const result = await patchStudy({
+    const result = await postStudy({
       form: formData,
-      studyId: id,
     });
     closeModal();
     const studyID = result.studyId;
