+++ src/features/edit-study/form/EditStudyForm.tsx
@@ -3,10 +3,10 @@ import { FormProvider, useForm } from "react-hook-form";
 
 import EditStudyForm1 from "@/features/edit-study/form/EditStudyForm1";
 
-import { StudyForm } from "@/entities/study/studies/api/studies.api.type";
+import { PatchStudyForm } from "@/entities/study/studies/api/studies.api.type";
 
 const EditStudyForm = () => {
-  const methods = useForm<StudyForm>({
+  const methods = useForm<PatchStudyForm>({
     mode: "onChange",
     defaultValues: {
       title: "",
@@ -19,9 +19,6 @@ const EditStudyForm = () => {
       schedule: "",
       content: "",
       requirements: "",
-      rules: [{ ruleNum: 1, content: "" }],
-      benefits: [{ benefitNum: 1, content: "" }],
-      applicationQuestions: [],
     },
   });
